# tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "actjs",
    "description": "Self-hosted actor framework HTTP API",
    "version": "0.3.0"
  },
...
  "paths": {
...
    "/v1/actors/{class}": {
      "post": {
        "summary": "Create an actor (mints a fresh id; onInit fires on first call)",
        "tags": [
          "actors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "def": {
                  "type": "unknown"
                },
                "type": "unknown"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "toJSONSchema",
            "required": true
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "class": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                }
              }
            },
            "in": "path",
            "name": "def",
            "required": true
          },
          {
            "schema": "object",
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "in": "path",
            "name": "parse",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParse",
            "required": true
          },
          {
            "in": "path",
            "name": "parseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "spa",
            "required": true
          },
          {
            "in": "path",
            "name": "encode",
            "required": true
          },
          {
            "in": "path",
            "name": "decode",
            "required": true
          },
          {
            "in": "path",
            "name": "encodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "decodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "extend",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "class": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "id": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "manifest": {
                        "def": {
                          "type": "optional",
                          "innerType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          }
                        },
                        "type": "optional"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actors/{class}/{id}": {
      "get": {
        "summary": "Read the current snapshot of an actor",
        "tags": [
          "actors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "toJSONSchema",
            "required": true
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "class": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                },
                "id": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": null,
                  "minLength": 1,
                  "maxLength": null
                }
              }
            },
            "in": "path",
            "name": "def",
            "required": true
          },
          {
            "schema": "object",
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "in": "path",
            "name": "parse",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParse",
            "required": true
          },
          {
            "in": "path",
            "name": "parseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "spa",
            "required": true
          },
          {
            "in": "path",
            "name": "encode",
            "required": true
          },
          {
            "in": "path",
            "name": "decode",
            "required": true
          },
          {
            "in": "path",
            "name": "encodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "decodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "extend",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "class": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "id": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "version": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "seq": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "state": {
                        "def": {
                          "type": "unknown"
                        },
                        "type": "unknown"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Tombstone an actor",
        "tags": [
          "actors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "toJSONSchema",
            "required": true
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "class": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                },
                "id": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": null,
                  "minLength": 1,
                  "maxLength": null
                }
              }
            },
            "in": "path",
            "name": "def",
            "required": true
          },
          {
            "schema": "object",
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "in": "path",
            "name": "parse",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParse",
            "required": true
          },
          {
            "in": "path",
            "name": "parseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "spa",
            "required": true
          },
          {
            "in": "path",
            "name": "encode",
            "required": true
          },
          {
            "in": "path",
            "name": "decode",
            "required": true
          },
          {
            "in": "path",
            "name": "encodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "decodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "extend",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "class": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "id": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "tombstoned": {
                        "def": {
                          "type": "literal",
                          "values": [
                            true
                          ]
                        },
                        "type": "literal",
                        "values": {}
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actors/{class}/{id}/{method}": {
      "post": {
        "summary": "Invoke a handler on an actor",
        "tags": [
          "actors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "def": {
                  "type": "unknown"
                },
                "type": "unknown"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "toJSONSchema",
            "required": true
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "class": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                },
                "id": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": null,
                  "minLength": 1,
                  "maxLength": null
                },
                "method": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                }
              }
            },
            "in": "path",
            "name": "def",
            "required": true
          },
          {
            "schema": "object",
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "in": "path",
            "name": "parse",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParse",
            "required": true
          },
          {
            "in": "path",
            "name": "parseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "spa",
            "required": true
          },
          {
            "in": "path",
            "name": "encode",
            "required": true
          },
          {
            "in": "path",
            "name": "decode",
            "required": true
          },
          {
            "in": "path",
            "name": "encodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "decodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecodeAsync",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "class": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "id": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "method": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "result": {
                        "def": {
                          "type": "unknown"
                        },
                        "type": "unknown"
                      },
                      "manifest": {
                        "def": {
                          "type": "optional",
                          "innerType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          }
                        },
                        "type": "optional"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}